+++ R/get_nested_model_coefficients.R
@@ -6,7 +6,9 @@ get_nested_model_coefficients <- function(model,
                                           train_data = NULL,
                                           test_data = NULL,
                                           fold_info = NULL,
-                                          include_fold_columns = TRUE) {
+                                          include_fold_columns = TRUE,
+                                          caller = "") {
+  # model <- NULL
   if (is.null(model)) {
     NA_coeffs <- tibble::tibble(
       "Fold Column" = NA,
@@ -29,20 +31,57 @@ get_nested_model_coefficients <- function(model,
     return(nested_NA_coeffs)
   }
 
-  coefs_tidy <- tryCatch(
-    {
-      # Custom tidy methods?
-      parameters::model_parameters(model) %>%
-        parameters::standardize_names(style = "broom") %>%
-        tibble::as_tibble()
-    },
+  coefs_tidy <- tryCatch({
+
+    # Extract the model parameters quietly
+    model_params_process <- run_quietly(
+      fn = function(model) {
+        parameters::model_parameters(model) %>%
+          parameters::standardize_names(style = "broom") %>%
+          tibble::as_tibble()
+      },
+      model = model,
+      msg_context_fn = function(m){m}
+    )
+
+    # Extract the parameters and side effects
+    model_params <- model_params_process[["result"]]
+    warnings <- model_params_process[["warnings"]]
+    messages <- model_params_process[["messages"]]
+
+    # Function for creating message with context
+    message_creator <- function(m){
+      create_message(
+        m = m,
+        caller = caller,
+        formula = NULL,
+        fold_col = fold_info[["fold_column"]],
+        fold = fold_info[["rel_fold"]]
+      )
+    }
+
+    # Message the caught messages to the user
+    throw_messages(messages = messages, msg_context_fn = message_creator)
+    # Throw the caught warnings
+    throw_warnings(warnings = warnings, msg_context_fn = message_creator)
+
+    # If we have warnings, we return an NA tibble
+    # TODO Is this behavior we want? Depends on potential warnings thrown by model_parameters()
+    # It is consistent with previous implementation
+    if (length(warnings) > 0) {
+      get_nested_model_coefficients(model = NULL, include_fold_columns = include_fold_columns)[[1]]
+    } else {
+      model_params
+    }
+  },
     error = function(e) {
       # If parameters::model_parameters wasn't implemented for the model type
       # let's grab the coefficients manually if possible
 
-      if (grepl("Error: Sorry, `model_parameters()` does currently not work for objects of class",
+      if (grepl("does currently not work for objects of class",
                 as.character(e),
                 ignore.case = TRUE)) {
+
         # Try to extract coefficients
         coefs <- tryCatch(
           {
@@ -72,12 +111,6 @@ get_nested_model_coefficients <- function(model,
           get_nested_model_coefficients(model = NULL, include_fold_columns = include_fold_columns)[[1]]
         )
       }
-    },
-    warning = function(w) {
-      warning(w)
-      return(
-        get_nested_model_coefficients(model = NULL, include_fold_columns = include_fold_columns)[[1]]
-      )
     }
   )
 
@@ -98,4 +131,5 @@ get_nested_model_coefficients <- function(model,
   coefs_tidy %>%
     nest_models() %>%
     dplyr::pull(.data$Coefficients)
+
 }