+++ tests/testthat/test_validate_fn.R
@@ -861,18 +861,16 @@ test_that("fuzz testing validate_fn()", {
   # Testing side effects
   # Assigning side effects
   side_effects_16417 <- xpectr::capture_side_effects(validate_fn(train_data = dat_ready, formulas = "diagnosis~score+(1|session)", type = "binomial", model_fn = glm_model_fn, predict_fn = glm_predict_fn, test_data = NULL, preprocess_fn = glm_preprocess_fn, preprocess_once = FALSE, hyperparameters = hparams, partitions_col = ".partitions", cutoff = 0.5, positive = 2, metrics = list(all = FALSE, Accuracy = TRUE, Sensitivity = TRUE), rm_nc = FALSE, parallel = FALSE, verbose = FALSE), reset_seed = TRUE)
-
-
   expect_equal(
     xpectr::strip(side_effects_16417[['warnings']]),
     xpectr::strip(c("validatefn prediction from a rankdeficient fit may be misleadingForFormula diagnosisscore1sessionFold column partitionsFold 2",
-                      "Model matrix is rank deficient Parameters 1 sessionTRUE were not estimable",
-                      "validatefn prediction from a rankdeficient fit may be misleadingForFormula diagnosisscore1sessionFold column partitionsFold 2",
-                      "Model matrix is rank deficient Parameters 1 sessionTRUE were not estimable",
-                      "validatefn prediction from a rankdeficient fit may be misleadingForFormula diagnosisscore1sessionFold column partitionsFold 2",
-                      "Model matrix is rank deficient Parameters 1 sessionTRUE were not estimable",
-                      "validatefn prediction from a rankdeficient fit may be misleadingForFormula diagnosisscore1sessionFold column partitionsFold 2",
-                      "Model matrix is rank deficient Parameters 1 sessionTRUE were not estimable"
+                    "validatefn Model matrix is rank deficient Parameters 1 sessionTRUE were not estimableForFold column partitionsFold 2",
+                    "validatefn prediction from a rankdeficient fit may be misleadingForFormula diagnosisscore1sessionFold column partitionsFold 2",
+                    "validatefn Model matrix is rank deficient Parameters 1 sessionTRUE were not estimableForFold column partitionsFold 2",
+                    "validatefn prediction from a rankdeficient fit may be misleadingForFormula diagnosisscore1sessionFold column partitionsFold 2",
+                    "validatefn Model matrix is rank deficient Parameters 1 sessionTRUE were not estimableForFold column partitionsFold 2",
+                    "validatefn prediction from a rankdeficient fit may be misleadingForFormula diagnosisscore1sessionFold column partitionsFold 2",
+                    "validatefn Model matrix is rank deficient Parameters 1 sessionTRUE were not estimableForFold column partitionsFold 2"
     )),
     fixed = TRUE)
   expect_equal(
@@ -2182,9 +2180,9 @@ test_that("fuzz testing gaussian lm model with validate_fn()",{
     xpectr::strip(side_effects_15190[['warnings']]),
     xpectr::strip(
       c("validatefn prediction from a rankdeficient fit may be misleadingForFormula score diagnosis 1sessionFold column partitionsFold 2",
-        "Model matrix is rank deficient Parameters 1 sessionTRUE were not estimable",
-        "validatefn prediction from a rankdeficient fit may be misleadingForFormula score diagnosis 1sessionFold column partitionsFold 2",
-        "Model matrix is rank deficient Parameters 1 sessionTRUE were not estimable"
+      "validatefn Model matrix is rank deficient Parameters 1 sessionTRUE were not estimableForFold column partitionsFold 2",
+      "validatefn prediction from a rankdeficient fit may be misleadingForFormula score diagnosis 1sessionFold column partitionsFold 2",
+      "validatefn Model matrix is rank deficient Parameters 1 sessionTRUE were not estimableForFold column partitionsFold 2"
       )),
     fixed = TRUE)
   expect_equal(
@@ -3193,7 +3191,7 @@ test_that("fuzz testing multinomial nnet model with validate_fn()", {
   side_effects_19346 <- xpectr::capture_side_effects(validate_fn(train_data = data_mc, formulas = "target ~ predictor_1 + predictor_2 + predictor_3", type = "gaussian", model_fn = multinom_model_fn, predict_fn = multinom_predict_fn, test_data = NULL, preprocess_fn = NULL, preprocess_once = FALSE, hyperparameters = hparams, partitions_col = ".partitions", metrics = list(), rm_nc = FALSE, parallel = FALSE, verbose = FALSE), reset_seed = TRUE)
   expect_equal(
     xpectr::strip(side_effects_19346[['error']]),
-    xpectr::strip("---\nvalidate_fn(): Error in run_predict_fn(test_data = test_data, train_data = train_data, : When 'type'/'family' is 'gaussian', the predictions must be a vector or matrix / data frame with one column but was a matrix with 3 columns. Did you specify 'predict_fn' correctly?\n\nFor:\nFormula: target ~ predictor_1 + predictor_2 + predictor_3\nFold column: .partitions\nFold: 2\n"),
+    xpectr::strip("---\nvalidate_fn(): Error in fn(...): When 'type'/'family' is 'gaussian', the predictions must be a vector or matrix / data frame with one column but was a matrix with 3 columns. Did you specify 'predict_fn' correctly?\n\nFor:\nFormula: target ~ predictor_1 + predictor_2 + predictor_3\nFold column: .partitions\nFold: 2\n"),
     fixed = TRUE)
   expect_equal(
     xpectr::strip(side_effects_19346[['error_class']]),